Compelto! Full. Com segurança e tudo.

+++ AppsAdmin/src/main/java/com/bergermobile/rest/controller/ContentCommandController.java
@@ -3,15 +3,11 @@ package com.bergermobile.rest.controller;
 import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
 import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.webmvc.RepositoryRestController;
 import org.springframework.hateoas.Resource;
-import org.springframework.hateoas.Resources;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -19,38 +15,20 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.bergermobile.persistence.domain.Content;
 import com.bergermobile.persistence.domain.Menu;
 import com.bergermobile.rest.domain.ContentRest;
 import com.bergermobile.rest.service.ContentService;
 
 @RepositoryRestController
 //@RestController
-public class ContentController {
+public class ContentCommandController {
 	
-	static Log LOG = LogFactory.getLog(ContentController.class);
+	static Log LOG = LogFactory.getLog(ContentCommandController.class);
 	
 	@Autowired
 	ContentService contentService;
 		
-	/**
-	 * @param appId
-	 * @param inLocale
-	 */
-	@PreAuthorize("hasAnyRole('ROLE_USER','ROLE_ADMIN')")
-	@RequestMapping(value = "/contents/full/app/{appId}", method = RequestMethod.GET)
-	@ResponseBody ResponseEntity<?> getContentsFull(@PathVariable String appId) {
-		LOG.debug("ContentController - full");
-		
-		List<Content> contents = new ArrayList<Content>();
-		Resources<Content> resources = new Resources<Content>(contents);
-		resources.add(linkTo(methodOn(ContentController.class).getContentsFull(appId)).withSelfRel()); 
-
-		return ResponseEntity.ok(resources);
-	}
-	
 	@RequestMapping(value = "/contents", method = RequestMethod.POST)
 	@PreAuthorize("hasAnyRole('ROLE_USER','ROLE_ADMIN')")
 	ResponseEntity<?> save(@RequestBody ContentRest contentRest) {
@@ -64,7 +42,7 @@ public class ContentController {
 		Menu menu = contentService.getMenuFromContent(contentRest);
 		
 		Resource<Menu> resource = new Resource<>(menu);
-		resource.add(linkTo(methodOn(ContentController.class).save(contentRest)).withSelfRel());
+		resource.add(linkTo(methodOn(ContentCommandController.class).save(contentRest)).withSelfRel());
 		
 		return ResponseEntity.ok(resource);
 	}
@@ -72,13 +50,16 @@ public class ContentController {
 	@RequestMapping(value = "/contents/groupId/{groupId}", method = RequestMethod.DELETE)
 	@PreAuthorize("hasAnyRole('ROLE_USER','ROLE_ADMIN')")
 	ResponseEntity<?> deleteByGroupId(@PathVariable String groupId) {
+		if (!contentService.isAuthorized(groupId)) {
+			throw new AccessDeniedException("User not authorized to delete this content");
+		}
 		Menu menu = contentService.getMenuFromGroupId(groupId);
 		
 		// invoke an asynchronous method to remove unused files
 		contentService.removeUnusedFiles();
 		
 		Resource<Menu> resource = new Resource<>(menu);
-		resource.add(linkTo(methodOn(ContentController.class).deleteByGroupId(groupId)).withSelfRel()); 
+		resource.add(linkTo(methodOn(ContentCommandController.class).deleteByGroupId(groupId)).withSelfRel()); 
 		
 		contentService.deleteByGroupId(groupId);
 		return ResponseEntity.ok(resource);
